* EOControl/EOCheapArray.m         * EOControl/EODebug.m         * EOControl/EODetailDataSource.m         * EOControl/EOEditingContext.m         * EOControl/EOFaultHandler.m         * EOControl/EOFault.m         * EOControl/EOGenericRecord.m         * EOControl/EOKeyComparisonQualifier.m         * EOControl/EOKeyValueArchiver.m         * EOControl/EOKeyValueCoding.m         * EOControl/EOKeyValueQualifier.m         * EOControl/EOMutableKnownKeyDictionary.m         * EOControl/EOObjectStoreCoordinator.m         * EOControl/EOPrivate.m         * EOControl/EOQualifier.m         * EOControl/EOSortOrdering.m         fix format string warnings and do not directly access isa
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gdl2/trunk@37737 72102866-910b-0410-8b05-ffd578937521

+++ EOControl/EOMutableKnownKeyDictionary.m
@@ -183,7 +183,7 @@ RCS_ID("$Id$")
 
   for (i = 0; i < _count; i++)
     {
-      dscr = [dscr stringByAppendingFormat: @"%@ [%d] ",
+      dscr = [dscr stringByAppendingFormat: @"%@ [%"PRIuPTR"] ",
                    _keys[i],
                    i];
     }
@@ -490,7 +490,7 @@ RCS_ID("$Id$")
   NSString *dscr;
   NSMutableString *offsets = [NSMutableString string];
   NSUInteger i;
-  int count = [_destinationDescription count];
+  NSUInteger count = [_destinationDescription count];
 
   dscr = [NSString stringWithFormat: @"<%s %p - ",
 		   object_getClassName(self),
@@ -501,7 +501,7 @@ RCS_ID("$Id$")
 	       [_destinationDescription description]];
 
   for (i = 0; i < count; i++)
-    [offsets appendFormat: @" %d", _sourceOffsetForDestinationOffset[i]];
+    [offsets appendFormat: @" %"PRIuPTR, _sourceOffsetForDestinationOffset[i]];
 
   dscr = [dscr stringByAppendingFormat:
 		 @"\nsourceOffsetForDestinationOffset:%@>", offsets];
